拆分重构 LocalizationCodeTransformer
